pyproject.toml
[tool.poetry]
name = "pathway_indexer"
version = "0.0.1"
description = "Create and maintain the index for the BYU Pathway service missionary chatbot"
authors = ["Dallan Quass <fdallan@gmail.com>"]
repository = "https://github.com/DallanQ/pathway-indexer"
documentation = "https://DallanQ.github.io/pathway-indexer/"
readme = "README.md"
packages = [
  {include = "pathway_indexer"}
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
beautifulsoup4 = "^4.12.3"
ipykernel = "^6.29.5"
requests = "^2.32.3"
unstructured-client = "^0.25.5"
llama-index = "^0.10.65"
llama-parse = "^0.4.9"
python-dotenv = "^1.0.1"
optuna = "^3.6.1"
chromadb = "^0.5.5"
llama-index-vector-stores-chroma = "^0.1.10"
optuna-dashboard = "^0.16.0"
pymilvus = "^2.4.5"
qdrant-client = "^1.11.0"
markdownify = "^0.13.1"
spacy = "^3.7.6"
python-frontmatter = "^1.1.0"
fuzzywuzzy = "^0.18.0"
playwright = "^1.46.0"
matplotlib = "^3.9.2"
voyageai = "^0.2.3"
llama-index-vector-stores-pinecone = "0.1.3"
nest-asyncio = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
tox = "^4.11.1"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["pathway_indexer"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pathway_indexer"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]
